refactor: moved cli commands into cli package
